requirements/dev.txt
-r base.txt
django-nose>=1.4.2,<2
beautifulsoup4==4.8.1
